this gets the fromerly private methods moved off the factory.
