+++ com.brgs.orm/Azure/AzureTableBuilder.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using com.brgs.orm.Azure.helpers;
+using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Table;
 
 namespace com.brgs.orm.Azure
@@ -9,15 +10,17 @@ namespace com.brgs.orm.Azure
     {
         private ICloudStorageAccount account { get; set; }
         private AzureFormatHelper helpers { get; set; }
+        private string Collection { get; set; }
         public AzureTableBuilder(ICloudStorageAccount acc)
         {
             account = acc;
             helpers = new AzureFormatHelper();
         }
-        public AzureTableBuilder(ICloudStorageAccount acc, string Collection)
+        public AzureTableBuilder(ICloudStorageAccount acc, string collection)
         {
             account = acc;
             helpers = new AzureFormatHelper();
+            Collection = collection;
             
         }
         public async Task<T> GetAsync<T>(TableQuery query, string collection)
@@ -48,6 +51,23 @@ namespace com.brgs.orm.Azure
             } while (token != null);
 
             return outVal;            
+        }
+        public async Task<TableResult> PostAsync<T>(T record)
+        {
+            try 
+            {                
+                var tableClient = account.CreateCloudTableClient();
+                var table = tableClient.GetTableReference(Collection);
+                bool complete = table.CreateIfNotExistsAsync().Result;
+                var insert = TableOperation.InsertOrMerge((ITableEntity) record);
+
+                var val =  await table.ExecuteAsync(insert);
+                return val;
+
+            } catch (StorageException e )
+            {
+                throw new Exception(e.RequestInformation.ExtendedErrorInformation.ToString());
+            }            
         }
     }
 }
